vacancy backend work with userNames

+++ api/interactors/vacancy.interactor.js
@@ -1,18 +1,39 @@
 const vacancyRepository = require('../infrastructure/vacancy.repository')
+const PermissionError = require('../exceptions/permission.error')
 
 async function getList() {
   return await vacancyRepository.getVacancies()
 }
 
-async function addVacancy(vacancy) {
+async function getUserVacancies(userName) {
+  return await vacancyRepository.getByAuthorUserName(userName)
+}
+
+async function addVacancy(userName, vacancy) {
+  vacancy.authorUserName = userName
+
   return await vacancyRepository.addVacancy(vacancy)
 }
 
-async function editVacancy(id, vacancy) {
-  return await vacancyRepository.editVacancy(vacancy)
+async function editVacancy(userName, id, vacancy) {
+  let foundVacancy = await vacancyRepository.getById(id)
+
+  if (foundVacancy.authorUserName !== userName) {
+    throw new PermissionError('edit is permitted only for author')
+  }
+
+  Object.assign(foundVacancy, vacancy)
+
+  return await vacancyRepository.editVacancy(foundVacancy)
 }
 
-async function deleteVacancy(id) {
+async function deleteVacancy(userName, id) {
+  let foundVacancy = await vacancyRepository.getById(id)
+
+  if (foundVacancy.authorUserName !== userName) {
+    throw new PermissionError('delete is permitted only for author')
+  }
+
   return await vacancyRepository.deleteVacancy(id)
 }
 
@@ -20,5 +41,6 @@ module.exports = {
   getList,
   addVacancy,
   editVacancy,
-  deleteVacancy
+  deleteVacancy,
+  getUserVacancies
 }
